Changed function W_LC::read_parameters, class LJ_Potential, wrote W_LC::maxwell_boltzmann();
The functions reads now thermostat options. If these options are not in the .parameter file, they are set by the constructor.
The class LJ_Potential has now a read_parameter function which read sigma, epsilon r_cut out of the paramaterfile.
The function maxwell_boltzmann initializes the velocities of all particles, if thermostat options are written in .paramter file.
The function is called in the W_LC::read_particles function before the particles are distributed over their cells.